Second optimization pass
